Image Captioning using Deep Learning/requirements.txt
tensorflow==2.18.0
keras==3.5.0 
numpy==1.26.4
gradio==5.29.0
pillow==11.1.0